refactor: remove inheritance between replay repositories; extract shared types
ReplaysRepository and ReplaysRepositoryV04 are now fully independent sealed classes. ReplayDb and StringArrayHandler are extracted to their own file so neither repository depends on the other.
